chore: rename packanalytics to just analytics
