+++ app/components/Analytics/ElevarEvents/events.ts
@@ -1,18 +1,16 @@
-import {v4 as uuidv4} from 'uuid';
-
-import {PackEventName} from '../constants';
+import {AnalyticsEvent} from '../constants';
 
 import {
   flattenConnection,
-  pathWithoutLocalePrefix,
   generateUserProperties,
   mapCartLine,
   mapProductItemVariant,
   mapProductPageVariant,
   mapProductItemProduct,
+  pathWithoutLocalePrefix,
 } from './utils';
 
-export const ANALYTICS_NAME = 'GA4Events';
+export const ANALYTICS_NAME = 'ElevarEvents';
 
 const PAGE_TYPES: Record<string, string> = {
   '/': 'home',
@@ -38,26 +36,26 @@ const PAGE_TYPES: Record<string, string> = {
 
 const logSubscription = ({
   data,
-  packEventName,
+  analyticsEvent,
 }: {
   data: Record<string, any>;
-  packEventName: string;
+  analyticsEvent: string;
 }) => {
   console.log(
-    `${ANALYTICS_NAME}: 📥 subscribed to analytics for \`${packEventName}\`:`,
+    `${ANALYTICS_NAME}: 📥 subscribed to analytics for \`${analyticsEvent}\`:`,
     data,
   );
 };
 
 const logError = ({
-  packEventName,
+  analyticsEvent,
   message = 'Unknown error',
 }: {
-  packEventName: string;
+  analyticsEvent: string;
   message?: string | unknown;
 }) => {
   console.error(
-    `${ANALYTICS_NAME}: ❌ error from \`${packEventName}\`: ${message}`,
+    `${ANALYTICS_NAME}: ❌ error from \`${analyticsEvent}\`: ${message}`,
   );
 };
 
@@ -71,25 +69,24 @@ const emitEvent = ({
   onEmit?: (event: Record<string, any>) => void;
 }) => {
   try {
-    const emittedEvent = {
-      ...event,
-      event_id: uuidv4(),
-      event_time: new Date().toISOString(),
-    } as Record<string, any>;
-    if (window.gtag) {
-      window.gtag('event', emittedEvent.event, emittedEvent);
-    } else {
-      throw new Error('`window.gtag` is not defined.');
+    if (event.event === 'dl_route_update') {
+      if (!window.ElevarInvalidateContext) return;
+      window.ElevarInvalidateContext();
+      if (typeof onEmit === 'function') onEmit(event);
+      if (debug) console.log('ElevarEvents: ⏩ pushed context');
+      return;
     }
+    window.ElevarDataLayer = window.ElevarDataLayer ?? [];
+    window.ElevarDataLayer.push(event);
     if (debug)
       console.log(
-        `${ANALYTICS_NAME}: 🚀 event emitted for \`${emittedEvent.event}\`:`,
-        emittedEvent,
+        `${ANALYTICS_NAME}: 🚀 event emitted for \`${event.event}\`:`,
+        event,
       );
-    if (typeof onEmit === 'function') onEmit(emittedEvent);
+    if (typeof onEmit === 'function') onEmit(event);
   } catch (error) {
     logError({
-      packEventName: 'emitEvent',
+      analyticsEvent: 'emitEvent',
       message: error instanceof Error ? error.message : error,
     });
   }
@@ -103,9 +100,9 @@ const customerEvent = ({
   onEmit?: (event: Record<string, any>) => void;
   debug?: boolean;
 }) => {
-  const packEventName = PackEventName.CUSTOMER;
+  const analyticsEvent = AnalyticsEvent.CUSTOMER;
   try {
-    if (debug) logSubscription({data, packEventName});
+    if (debug) logSubscription({data, analyticsEvent});
 
     const {customer: providerCustomer, shop} = data;
     const {customer: customCustomer, cart} = data.customData;
@@ -123,21 +120,21 @@ const customerEvent = ({
       (previousPath?.startsWith('/collections') && previousPath) ||
       '';
     const event = {
-      event: 'user_data',
+      event: 'dl_user_data',
       user_properties: generateUserProperties({customer}),
+      cart_total: cart?.cost?.totalAmount?.amount || '0.0',
       ecommerce: {
-        currency_code: cart?.cost?.totalAmount?.currencyCode || shop?.currency,
+        currencyCode: cart?.cost?.totalAmount?.currencyCode || shop?.currency,
         cart_contents: {
           products:
             flattenConnection(cart?.lines)?.map(mapCartLine(list)) || [],
         },
-        cart_total: cart?.cost?.totalAmount?.amount || '0.0',
       },
     };
     emitEvent({event, onEmit, debug});
   } catch (error) {
     logError({
-      packEventName,
+      analyticsEvent,
       message: error instanceof Error ? error.message : error,
     });
   }
@@ -147,9 +144,9 @@ const viewPageEvent = ({
   debug,
   ...data
 }: Record<string, any> & {debug?: boolean}) => {
-  const packEventName = PackEventName.PAGE_VIEWED;
+  const analyticsEvent = AnalyticsEvent.PAGE_VIEWED;
   try {
-    if (debug) logSubscription({data, packEventName});
+    if (debug) logSubscription({data, analyticsEvent});
 
     const {url, customer} = data;
     if (!url) throw new Error('`url` parameter is missing.');
@@ -162,7 +159,7 @@ const viewPageEvent = ({
         PAGE_TYPES[pathname.split('/').slice(0, -1).join('/')] ||
         '';
     const event = {
-      event: 'route_update',
+      event: 'dl_route_update',
       user_properties: generateUserProperties({customer}),
       page: {
         path: pathname,
@@ -174,7 +171,7 @@ const viewPageEvent = ({
     emitEvent({event, debug});
   } catch (error) {
     logError({
-      packEventName,
+      analyticsEvent,
       message: error instanceof Error ? error.message : error,
     });
   }
@@ -184,9 +181,9 @@ const viewProductEvent = ({
   debug,
   ...data
 }: Record<string, any> & {debug?: boolean}) => {
-  const packEventName = PackEventName.PRODUCT_VIEWED;
+  const analyticsEvent = AnalyticsEvent.PRODUCT_VIEWED;
   try {
-    if (debug) logSubscription({data, packEventName});
+    if (debug) logSubscription({data, analyticsEvent});
 
     const {customer, shop} = data;
     const {product, selectedVariant} = data.customData;
@@ -208,10 +205,10 @@ const viewProductEvent = ({
     const previousPath = sessionStorage.getItem('PREVIOUS_PATH');
     const list = previousPath?.startsWith('/collections') ? previousPath : '';
     const event = {
-      event: 'view_item',
+      event: 'dl_view_item',
       user_properties: generateUserProperties({customer}),
       ecommerce: {
-        currency_code: variant.price?.currencyCode || shop?.currency,
+        currencyCode: variant.price?.currencyCode || shop?.currency,
         detail: {
           actionField: {list, action: 'detail'},
           products: [variant].map(mapProductPageVariant(list)),
@@ -221,7 +218,7 @@ const viewProductEvent = ({
     emitEvent({event, debug});
   } catch (error) {
     logError({
-      packEventName,
+      analyticsEvent,
       message: error instanceof Error ? error.message : error,
     });
   }
@@ -231,9 +228,9 @@ const viewCollectionEvent = ({
   debug,
   ...data
 }: Record<string, any> & {debug?: boolean}) => {
-  const packEventName = PackEventName.COLLECTION_VIEWED;
+  const analyticsEvent = AnalyticsEvent.COLLECTION_VIEWED;
   try {
-    if (debug) logSubscription({data, packEventName});
+    if (debug) logSubscription({data, analyticsEvent});
 
     const {customer, shop} = data;
     const {collection} = data.customData;
@@ -246,21 +243,21 @@ const viewCollectionEvent = ({
       : '';
     const products = flattenConnection(collection.products);
     const event = {
-      event: 'view_item_list',
+      event: 'dl_view_item_list',
       user_properties: generateUserProperties({customer}),
       ecommerce: {
-        currency_code:
+        currencyCode:
           flattenConnection(products?.[0]?.variants)?.[0]?.price
             ?.currencyCode || shop?.currency,
         collection_title: collection.title,
         collection_id: collection.id?.split('/').pop(),
-        products: products?.slice(0, 12).map(mapProductItemProduct(list)),
+        impressions: products?.slice(0, 7).map(mapProductItemProduct(list)),
       },
     };
     emitEvent({event, debug});
   } catch (error) {
     logError({
-      packEventName,
+      analyticsEvent,
       message: error instanceof Error ? error.message : error,
     });
   }
@@ -270,9 +267,9 @@ const viewSearchResultsEvent = ({
   debug,
   ...data
 }: Record<string, any> & {debug?: boolean}) => {
-  const packEventName = PackEventName.SEARCH_VIEWED;
+  const analyticsEvent = AnalyticsEvent.SEARCH_VIEWED;
   try {
-    if (debug) logSubscription({data, packEventName});
+    if (debug) logSubscription({data, analyticsEvent});
 
     const {searchResults, searchTerm, customer, shop} = data;
     if (!searchResults || !searchTerm)
@@ -281,23 +278,29 @@ const viewSearchResultsEvent = ({
       );
 
     const event = {
-      event: 'view_search_results',
+      event: 'dl_view_search_results',
       user_properties: generateUserProperties({customer}),
       ecommerce: {
-        currency_code:
+        currencyCode:
           flattenConnection(searchResults[0]?.variants)?.[0]?.price
             ?.currencyCode || shop?.currency,
         actionField: {
           list: 'search results',
           search_term: searchTerm,
         },
-        products: searchResults.slice(0, 12).map(mapProductItemProduct()),
+        impressions: searchResults.slice(0, 7).map(mapProductItemProduct()),
       },
     };
-    emitEvent({event, debug});
+    // Elevar requires a `dl_user_data` event before `dl_view_search_results`
+    customerEvent({
+      ...data,
+      customer: customer || null,
+      onEmit: () => emitEvent({event, debug}),
+      debug,
+    });
   } catch (error) {
     logError({
-      packEventName,
+      analyticsEvent,
       message: error instanceof Error ? error.message : error,
     });
   }
@@ -307,9 +310,9 @@ const viewCartEvent = ({
   debug,
   ...data
 }: Record<string, any> & {debug?: boolean}) => {
-  const packEventName = PackEventName.CART_VIEWED;
+  const analyticsEvent = AnalyticsEvent.CART_VIEWED;
   try {
-    if (debug) logSubscription({data, packEventName});
+    if (debug) logSubscription({data, analyticsEvent});
 
     const {cart: providerCart, customer, shop, url} = data;
     const {cart: customCart} = data.customData;
@@ -329,23 +332,27 @@ const viewCartEvent = ({
       (previousPath?.startsWith('/collections') && previousPath) ||
       '';
     const event = {
-      event: 'view_cart',
+      event: 'dl_view_cart',
       user_properties: generateUserProperties({customer}),
+      cart_total: cart?.cost?.totalAmount?.amount || '0.0',
       ecommerce: {
-        currency_code: cart?.cost?.totalAmount?.currencyCode || shop?.currency,
+        currencyCode: cart?.cost?.totalAmount?.currencyCode || shop?.currency,
         actionField: {list: 'Shopping Cart'},
-        products:
-          flattenConnection(cart?.lines)?.slice(0, 12).map(mapCartLine(list)) ||
+        impressions:
+          flattenConnection(cart?.lines)?.slice(0, 7).map(mapCartLine(list)) ||
           [],
-        cart_id: cart?.id?.split('/').pop() || 'uninitialized',
-        cart_total: cart?.cost?.totalAmount?.amount || '0.0',
-        cart_count: cart?.totalQuantity || 0,
       },
     };
-    emitEvent({event, debug});
+    // Elevar requires a `dl_user_data` event before `dl_view_cart`
+    customerEvent({
+      ...data,
+      customer: customer || null,
+      onEmit: () => emitEvent({event, debug}),
+      debug,
+    });
   } catch (error) {
     logError({
-      packEventName,
+      analyticsEvent,
       message: error instanceof Error ? error.message : error,
     });
   }
@@ -355,9 +362,9 @@ const addToCartEvent = ({
   debug,
   ...data
 }: Record<string, any> & {debug?: boolean}) => {
-  const packEventName = PackEventName.PRODUCT_ADD_TO_CART;
+  const analyticsEvent = AnalyticsEvent.PRODUCT_ADD_TO_CART;
   try {
-    if (debug) logSubscription({data, packEventName});
+    if (debug) logSubscription({data, analyticsEvent});
 
     const {cart, currentLine, customer, shop} = data;
     if (!cart || !currentLine)
@@ -370,23 +377,21 @@ const addToCartEvent = ({
       (previousPath?.startsWith('/collections') && previousPath) ||
       '';
     const event = {
-      event: 'add_to_cart',
+      event: 'dl_add_to_cart',
       user_properties: generateUserProperties({customer}),
+      cart_total: cart.cost?.totalAmount?.amount || '0.0',
       ecommerce: {
-        currency_code: cart.cost?.totalAmount?.currencyCode || shop?.currency,
+        currencyCode: cart.cost?.totalAmount?.currencyCode || shop?.currency,
         add: {
           actionField: {list},
           products: [currentLine].map(mapCartLine(list)),
         },
-        cart_id: cart.id?.split('/').pop(),
-        cart_total: cart.cost?.totalAmount?.amount || '0.0',
-        cart_count: cart.totalQuantity,
       },
     };
     emitEvent({event, debug});
   } catch (error) {
     logError({
-      packEventName,
+      analyticsEvent,
       message: error instanceof Error ? error.message : error,
     });
   }
@@ -396,9 +401,9 @@ const removeFromCartEvent = ({
   debug,
   ...data
 }: Record<string, any> & {debug?: boolean}) => {
-  const packEventName = PackEventName.PRODUCT_REMOVED_FROM_CART;
+  const analyticsEvent = AnalyticsEvent.PRODUCT_REMOVED_FROM_CART;
   try {
-    if (debug) logSubscription({data, packEventName});
+    if (debug) logSubscription({data, analyticsEvent});
 
     const {cart, prevLine, customer, shop} = data;
     if (!cart || !prevLine)
@@ -411,23 +416,21 @@ const removeFromCartEvent = ({
       (previousPath?.startsWith('/collections') && previousPath) ||
       '';
     const event = {
-      event: 'remove_from_cart',
+      event: 'dl_remove_from_cart',
       user_properties: generateUserProperties({customer}),
+      cart_total: cart.cost?.totalAmount?.amount || '0.0',
       ecommerce: {
-        currency_code: cart.cost?.totalAmount?.currencyCode || shop?.currency,
+        currencyCode: cart.cost?.totalAmount?.currencyCode || shop?.currency,
         remove: {
           actionField: {list},
           products: [prevLine].map(mapCartLine(list)),
         },
-        cart_id: cart.id?.split('/').pop(),
-        cart_total: cart.cost?.totalAmount?.amount || '0.0',
-        cart_count: cart.totalQuantity,
       },
     };
     emitEvent({event, debug});
   } catch (error) {
     logError({
-      packEventName,
+      analyticsEvent,
       message: error instanceof Error ? error.message : error,
     });
   }
@@ -437,9 +440,9 @@ const clickProductItemEvent = ({
   debug,
   ...data
 }: Record<string, any> & {debug?: boolean}) => {
-  const packEventName = PackEventName.PRODUCT_ITEM_CLICKED;
+  const analyticsEvent = AnalyticsEvent.PRODUCT_ITEM_CLICKED;
   try {
-    if (debug) logSubscription({data, packEventName});
+    if (debug) logSubscription({data, analyticsEvent});
 
     const {product, listIndex, searchTerm, customer, shop} = data;
     let {selectedVariant} = data;
@@ -465,10 +468,10 @@ const clickProductItemEvent = ({
     };
 
     const event = {
-      event: 'select_item',
+      event: 'dl_select_item',
       user_properties: generateUserProperties({customer}),
       ecommerce: {
-        currency_code: variant.price?.currencyCode || shop?.currency,
+        currencyCode: variant.price?.currencyCode || shop?.currency,
         click: {
           actionField: {
             list: searchTerm ? 'search results' : list,
@@ -482,59 +485,7 @@ const clickProductItemEvent = ({
     emitEvent({event, debug});
   } catch (error) {
     logError({
-      packEventName,
-      message: error instanceof Error ? error.message : error,
-    });
-  }
-};
-
-const clickProductVariantEvent = ({
-  debug,
-  ...data
-}: Record<string, any> & {debug?: boolean}) => {
-  const packEventName = PackEventName.PRODUCT_VARIANT_SELECTED;
-  try {
-    if (debug) logSubscription({data, packEventName});
-
-    const {
-      selectedVariant,
-      optionName,
-      optionValue,
-      fromGrouping,
-      customer,
-      shop,
-    } = data;
-    if (!selectedVariant)
-      throw new Error('`selectedVariant` parameter is missing.');
-
-    const list = window.location.pathname.startsWith('/collections')
-      ? window.location.pathname
-      : '';
-
-    const variant = {
-      ...selectedVariant,
-      image: selectedVariant.image || '',
-      list,
-    };
-
-    const event = {
-      event: 'dl_select_item_variant',
-      user_properties: generateUserProperties({customer}),
-      ecommerce: {
-        currency_code: variant.price?.currencyCode || shop?.currency,
-        click: {
-          actionField: {list, action: 'click'},
-          products: [variant].map(mapProductItemVariant(list)),
-          optionName,
-          optionValue: optionValue?.name,
-          fromGrouping: fromGrouping || false,
-        },
-      },
-    };
-    emitEvent({event, debug});
-  } catch (error) {
-    logError({
-      packEventName,
+      analyticsEvent,
       message: error instanceof Error ? error.message : error,
     });
   }
@@ -544,21 +495,21 @@ const customerLogInEvent = ({
   debug,
   ...data
 }: Record<string, any> & {debug?: boolean}) => {
-  const packEventName = PackEventName.CUSTOMER_LOGGED_IN;
+  const analyticsEvent = AnalyticsEvent.CUSTOMER_LOGGED_IN;
   try {
-    if (debug) logSubscription({data, packEventName});
+    if (debug) logSubscription({data, analyticsEvent});
 
     const {customer} = data;
     if (!customer) throw new Error('`customer` parameter is missing.');
 
     const event = {
-      event: 'login',
+      event: 'dl_login',
       user_properties: generateUserProperties({customer}),
     };
     emitEvent({event, debug});
   } catch (error) {
     logError({
-      packEventName,
+      analyticsEvent,
       message: error instanceof Error ? error.message : error,
     });
   }
@@ -568,21 +519,21 @@ const customerRegisterEvent = ({
   debug,
   ...data
 }: Record<string, any> & {debug?: boolean}) => {
-  const packEventName = PackEventName.CUSTOMER_REGISTERED;
+  const analyticsEvent = AnalyticsEvent.CUSTOMER_REGISTERED;
   try {
-    if (debug) logSubscription({data, packEventName});
+    if (debug) logSubscription({data, analyticsEvent});
 
     const {customer} = data;
     if (!customer) throw new Error('`customer` parameter is missing.');
 
     const event = {
-      event: 'sign_up',
+      event: 'dl_sign_up',
       user_properties: generateUserProperties({customer}),
     };
     emitEvent({event, debug});
   } catch (error) {
     logError({
-      packEventName,
+      analyticsEvent,
       message: error instanceof Error ? error.message : error,
     });
   }
@@ -592,9 +543,9 @@ const customerSubscribeEvent = ({
   debug,
   ...data
 }: Record<string, any> & {debug?: boolean}) => {
-  const packEventName = PackEventName.CUSTOMER_SUBSCRIBED;
+  const analyticsEvent = AnalyticsEvent.CUSTOMER_SUBSCRIBED;
   try {
-    if (debug) logSubscription({data, packEventName});
+    if (debug) logSubscription({data, analyticsEvent});
 
     const {email, phone} = data;
     if (!email && !phone)
@@ -602,7 +553,7 @@ const customerSubscribeEvent = ({
 
     if (email) {
       const event = {
-        event: 'subscribe',
+        event: 'dl_subscribe',
         lead_type: 'email',
         user_properties: {customer_email: email},
       };
@@ -610,7 +561,7 @@ const customerSubscribeEvent = ({
     }
     if (phone) {
       const event = {
-        event: 'subscribe',
+        event: 'dl_subscribe',
         lead_type: 'phone',
         user_properties: {customer_phone: phone},
       };
@@ -618,7 +569,7 @@ const customerSubscribeEvent = ({
     }
   } catch (error) {
     logError({
-      packEventName,
+      analyticsEvent,
       message: error instanceof Error ? error.message : error,
     });
   }
@@ -627,7 +578,6 @@ const customerSubscribeEvent = ({
 export {
   addToCartEvent,
   clickProductItemEvent,
-  clickProductVariantEvent,
   customerEvent,
   customerLogInEvent,
   customerRegisterEvent,
