+++ app/components/Analytics/GA4Events/GA4Events.tsx
@@ -1,6 +1,6 @@
 import {useEffect, useState} from 'react';
 
-import {PackEventName} from '../constants';
+import {AnalyticsEvent} from '../constants';
 
 import {
   viewPageEvent,
@@ -90,43 +90,43 @@ export function GA4Events({
     }
     /* register analytics events only until script is ready */
     if (!scriptLoaded) return;
-    subscribe(PackEventName.PAGE_VIEWED, (data: Data) => {
+    subscribe(AnalyticsEvent.PAGE_VIEWED, (data: Data) => {
       viewPageEvent({...data, customer, debug});
     });
-    subscribe(PackEventName.PRODUCT_VIEWED, (data: Data) => {
+    subscribe(AnalyticsEvent.PRODUCT_VIEWED, (data: Data) => {
       viewProductEvent({...data, customer, debug});
     });
-    subscribe(PackEventName.COLLECTION_VIEWED, (data: Data) => {
+    subscribe(AnalyticsEvent.COLLECTION_VIEWED, (data: Data) => {
       viewCollectionEvent({...data, customer, debug});
     });
-    subscribe(PackEventName.SEARCH_VIEWED, (data: Data) => {
+    subscribe(AnalyticsEvent.SEARCH_VIEWED, (data: Data) => {
       viewSearchResultsEvent({...data, customer, debug});
     });
-    subscribe(PackEventName.CART_VIEWED, (data: Data) => {
+    subscribe(AnalyticsEvent.CART_VIEWED, (data: Data) => {
       viewCartEvent({...data, customer, debug});
     });
-    subscribe(PackEventName.PRODUCT_VARIANT_SELECTED, (data: Data) => {
+    subscribe(AnalyticsEvent.PRODUCT_VARIANT_SELECTED, (data: Data) => {
       clickProductVariantEvent({...data, customer, debug});
     });
-    subscribe(PackEventName.PRODUCT_ITEM_CLICKED, (data: Data) => {
+    subscribe(AnalyticsEvent.PRODUCT_ITEM_CLICKED, (data: Data) => {
       clickProductItemEvent({...data, customer, debug});
     });
-    subscribe(PackEventName.PRODUCT_ADD_TO_CART, (data: Data) => {
+    subscribe(AnalyticsEvent.PRODUCT_ADD_TO_CART, (data: Data) => {
       addToCartEvent({...data, customer, debug});
     });
-    subscribe(PackEventName.PRODUCT_REMOVED_FROM_CART, (data: Data) => {
+    subscribe(AnalyticsEvent.PRODUCT_REMOVED_FROM_CART, (data: Data) => {
       removeFromCartEvent({...data, customer, debug});
     });
-    subscribe(PackEventName.CUSTOMER, (data: Data) => {
+    subscribe(AnalyticsEvent.CUSTOMER, (data: Data) => {
       customerEvent({...data, debug});
     });
-    subscribe(PackEventName.CUSTOMER_LOGGED_IN, (data: Data) => {
+    subscribe(AnalyticsEvent.CUSTOMER_LOGGED_IN, (data: Data) => {
       customerLogInEvent({...data, debug});
     });
-    subscribe(PackEventName.CUSTOMER_REGISTERED, (data: Data) => {
+    subscribe(AnalyticsEvent.CUSTOMER_REGISTERED, (data: Data) => {
       customerRegisterEvent({...data, debug});
     });
-    subscribe(PackEventName.CUSTOMER_SUBSCRIBED, (data: Data) => {
+    subscribe(AnalyticsEvent.CUSTOMER_SUBSCRIBED, (data: Data) => {
       customerSubscribeEvent({...data, debug});
     });
     ready();
