+++ app/components/Analytics/MetaPixelEvents/MetaPixelEvents.tsx
@@ -1,6 +1,6 @@
 import {useEffect, useState} from 'react';
 
-import {PackEventName} from '../constants';
+import {AnalyticsEvent} from '../constants';
 
 import {viewProductEvent, addToCartEvent, ANALYTICS_NAME} from './events';
 
@@ -69,10 +69,10 @@ export function MetaPixelEvents({
     }
     /* register analytics events only until script is ready */
     if (!scriptLoaded) return;
-    subscribe(PackEventName.PRODUCT_VIEWED, (data: Data) => {
+    subscribe(AnalyticsEvent.PRODUCT_VIEWED, (data: Data) => {
       viewProductEvent({...data, customer, debug});
     });
-    subscribe(PackEventName.PRODUCT_ADD_TO_CART, (data: Data) => {
+    subscribe(AnalyticsEvent.PRODUCT_ADD_TO_CART, (data: Data) => {
       addToCartEvent({...data, customer, debug});
     });
     ready();
