+++ app/components/Analytics/constants.ts
@@ -1,6 +1,6 @@
-import {AnalyticsEvent} from '@shopify/hydrogen';
+import {AnalyticsEvent as HydrogenAnalyticsEvent} from '@shopify/hydrogen';
 
-// AnalyticsEvent ------------------------------
+// Default AnalyticsEvent ------------------------------
 // PAGE_VIEWED: "page_viewed";
 // PRODUCT_VIEWED: "product_viewed";
 // COLLECTION_VIEWED: "collection_viewed";
@@ -11,8 +11,8 @@ import {AnalyticsEvent} from '@shopify/hydrogen';
 // PRODUCT_REMOVED_FROM_CART: "product_removed_from_cart";
 // CUSTOM_EVENT: `custom_${string}`;
 
-export const PackEventName = {
-  ...AnalyticsEvent,
+export const AnalyticsEvent = {
+  ...HydrogenAnalyticsEvent,
   PRODUCT_VARIANT_SELECTED: 'custom_product_variant_selected',
   PRODUCT_ITEM_CLICKED: 'custom_product_item_clicked',
   PRODUCT_QUICK_SHOP_VIEWED: 'custom_product_quick_shop_viewed',
@@ -20,7 +20,7 @@ export const PackEventName = {
   CUSTOMER_SUBSCRIBED: 'custom_customer_subscribed',
   CUSTOMER_LOGGED_IN: 'custom_customer_logged_in',
   CUSTOMER_REGISTERED: 'custom_customer_registered',
-} as typeof AnalyticsEvent & {
+} as typeof HydrogenAnalyticsEvent & {
   PRODUCT_VARIANT_SELECTED: 'custom_product_variant_selected';
   PRODUCT_ITEM_CLICKED: 'custom_product_item_clicked';
   PRODUCT_QUICK_SHOP_VIEWED: 'custom_product_quick_shop_viewed';
