+++ app/components/Analytics/index.ts
@@ -1,0 +1,1 @@
+export {Analytics} from './Analytics';
