--- app/components/PackAnalytics/index.ts
@@ -1,1 +1,0 @@
-export {PackAnalytics} from './PackAnalytics';
